requirements.txt
pandas>=2.0.0
neo4j>=5.14.0
python-dotenv>=1.0.0
flask>=3.0.0
ollama>=0.3.0
